pyproject.toml
[project]
name = "pubmed-paper-fetcher"
version = "0.1.0"
description = "Fetches and filters PubMed research papers"
authors = [
    {name = "Pravallika",email = "slpravallika7@gmail.com"}
]
license="MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "typer[all] (>=0.15.2,<0.16.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "biopython (>=1.85,<2.0)",
    "argparse (>=1.4.0,<2.0.0)",
    "pymupdf (>=1.25.4,<2.0.0)"
]

[tool.poetry.scripts]
get-papers-list = "pubmed_paper_fetcher.cli:main"


[tool.poetry]
packages = [{include = "pubmed_paper_fetcher", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
